[问题/需求描述]:行情数据更新，最快1s的时间间隔有点低 [解决方法]:支持ms级别的时间间隔 [影响范围]:策略数据获取时间间隔 [需求编号]:无

+++ interface/strategyEvent/strategyEvent.cpp
@@ -85,7 +85,7 @@ void StrategyEvent::TickSubscribeReqHandle(MsgStruct& msg)
     auto& marketSer = MarketService::getInstance();
     marketSer.ROLE(publishData).buildInstrumentList(insVec);
     marketSer.ROLE(publishData).buildKeywordList(keywordVec);
-    marketSer.ROLE(publishData).setInterval(std::atoi(reqInfo.interval().c_str()));
+    marketSer.ROLE(publishData).setInterval(float(std::atof(reqInfo.interval().c_str())));
 
     if (marketSer.ROLE(Market).ROLE(CtpMarketApi).marketApi != nullptr)
     {
